updated manifestProvider location
