+++ app/pages/sections/catalogue-solutions/order-item/manifestProvider.js
@@ -1,1 +1,7 @@
-// function to get the manifest based upon the type and provisioningType
+import fs from 'fs';
+import path from 'path';
+
+export const getSelectedPriceManifest = (provisioningType) => {
+  const rawManifest = fs.readFileSync(path.join(__dirname, `/${provisioningType.toLowerCase()}/manifest.json`));
+  return JSON.parse(rawManifest);
+};
